refactor: update bucket_schema client to use embedded fields
